api/media-service/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==10.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-flask==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
